% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_labels.R
\name{addSeriesBand}
\alias{addSeriesBand}
\title{addSeriesBand}
\usage{
addSeriesBand(
  dataFrame = NULL,
  seriesBandLower = NULL,
  seriesBandUpper = NULL,
  seriesScale = 1,
  colorBandSeries = NULL,
  borderBand = NULL,
  startDate = NULL,
  endDate = NULL,
  famePath = NULL
)
}
\description{
This plots the series band if supplied.
}
