% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barChart.R
\name{barChart}
\alias{barChart}
\title{Bar Chart}
\usage{
barChart(
  stacked = F,
  famePath = NULL,
  famePath2 = famePath,
  famePath3 = famePath,
  famePathLine = famePath,
  famePathLine2 = famePathLine,
  famePathLine3 = famePathLine,
  dataFrame = NULL,
  dataFrame2 = NULL,
  dataFrame3 = NULL,
  dataFrameLeft = dataFrame,
  dataFrameLeft2 = dataFrame2,
  dataFrameLeft3 = dataFrame3,
  dataFrameLine = dataFrame,
  dataFrameLine2 = dataFrame2,
  dataFrameLine3 = dataFrame3,
  seriesLine2 = NULL,
  seriesLine3 = NULL,
  chartTitle = "",
  chartTitleFontSize = opt$paneltitlecex,
  series = NULL,
  series2 = NULL,
  series3 = NULL,
  yMin = NULL,
  yMax = NULL,
  yStep = NULL,
  seriesScale = 1,
  unitsLabel = "",
  axisColorRight = "black",
  seriesLine = NULL,
  yMinLeft = NULL,
  yMaxLeft = NULL,
  yStepLeft = NULL,
  unitsLabelLeft = "",
  labelFontSize = opt$legend.cex,
  seriesScaleLeft = 1,
  seriesScaleLine = 1,
  seriesScale2 = 1,
  seriesScale3 = 1,
  seriesScaleLine2 = 1,
  seriesScaleLine3 = 1,
  axisColorLeft = "black",
  xAxisFontSize = opt$axis.cex,
  yAxisFontSize = opt$axis.cex,
  colorBandSeries = NA,
  seriesBandUpper = NA,
  seriesBandLower = NA,
  borderBand = F,
  startDate = NULL,
  endDate = NULL,
  startDate2 = NULL,
  endDate2 = NULL,
  startDate3 = NULL,
  endDate3 = NULL,
  xMajorTickFreq = NA,
  xMinorTickFreq = NA,
  xLabelFreq = "year",
  extraLimX = 0,
  legend = TRUE,
  legendNames = NULL,
  legendCol = 1,
  legendColPlacement = NULL,
  legendLocationX = 0.05,
  legendLocationY = 0.97,
  legendFontSize = opt$key.cex,
  frequencyLabel = "",
  footNote = "",
  footNotePlacement = NULL,
  footNoteFontSize = opt$foot.cex,
  lineType = NULL,
  lineWidth = NULL,
  colors = NULL,
  colorsLine = NULL,
  dynamicColors = FALSE,
  lineAtZero = T,
  printLastValue = TRUE,
  lastValueFreq = NA,
  lastValue = NULL,
  recessions = FALSE,
  hadj = opt$yaxis.pos,
  interp = FALSE,
  roundDig = 2,
  roundDigLeft = 2,
  customShading = FALSE,
  shadeStart = "",
  shadeEnd = "",
  palette = NULL,
  scheme = NULL,
  xBarTickAdj = 0.5,
  xBarTicksSeries1 = T,
  xBarTicksSeries2 = T,
  xBarTicksSeries3 = T,
  xTicksBelowAxis = T,
  barSpace = T,
  barBorder = T,
  barBorderColor = "black",
  barBorderWidth = 1,
  xSeriesDivider = T,
  series2LabelFreq = NA,
  series3LabelFreq = NA,
  barLabelFontSize = 0.5,
  monthLabelLetters = F,
  lastValuePrelim = F,
  lastValueLocationX = 0.9,
  lastValueLocationY = 0.9,
  dynamicDates = F,
  fullYearAsQuarters = T,
  PRESET = list(),
  dataType = NULL,
  dataTypeLine = NULL,
  series2MonthLabel = NULL,
  series2QuarterLabel = NULL,
  series3MonthLabel = NULL,
  series3QuarterLabel = NULL,
  tickMarksSeries1 = NULL,
  tickMarksSeries2 = NULL,
  tickMarksSeries3 = NULL,
  printLastMonthLabel = NULL,
  printLastMonthDayLabel = NULL,
  printLastQuarterLabel = NULL,
  lastValueLabelVerticalShift = NULL,
  lastValueLabelHorizontalShift = NULL,
  xLabelStop = NULL,
  xLabelYearUpperTicks = NULL,
  xLabelYearLowerTicks = NULL,
  xLabelShift = NULL
)
}
\description{
This creates a bar chart.
}
