% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_labels.R
\name{barChartLabels}
\alias{barChartLabels}
\title{defineXAxisBar}
\usage{
barChartLabels(
  bars,
  df = plot.data.full,
  chartType,
  series,
  series1Length,
  series2Length,
  series3Length,
  series2LabelFreq,
  series3LabelFreq,
  lastValuePrelim,
  chartParametersRight,
  barLabelFontSize,
  monthLabelLetters
)
}
\description{
This defines the X-axis for Bar Charts
}
