% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_labels.R
\name{chartLabels}
\alias{chartLabels}
\title{chartLabels}
\usage{
chartLabels(
  chartType,
  df = NULL,
  chartTitle = NULL,
  chartTitleFontSize = NULL,
  series = NULL,
  seriesLeft = NULL,
  legend = NULL,
  pointType = NULL,
  lineType = NULL,
  lineWidth = NULL,
  legendCol = NULL,
  legendLocationX = NULL,
  legendLocationY = NULL,
  legendFontSize = NULL,
  legendColPlacement = NULL,
  legendNames = NULL,
  footNote = NULL,
  footNoteFontSize = NULL,
  footNotePlacement = NULL,
  xLabelFreq,
  xTicksBelowAxis,
  unitsLabel = NULL,
  colors = NULL,
  unitsLabelLeft = NULL,
  frequencyLabel = NULL,
  axisColorRight = NULL,
  axisColorLeft = NULL,
  printLastValue = NULL,
  lastValueFreq = NULL,
  lastValue = NULL,
  lastValueLocationX,
  lastValueLocationY,
  lastValuePrelim = NULL,
  yMax = NULL,
  yMin = NULL,
  yStep = NULL,
  yMaxLeft = NULL,
  yMinLeft = NULL,
  yStepLeft = NULL,
  labelFontSize = NULL,
  seriesXLabel = ""
)
}
\description{
This creates the legend and title as well as extra labels such as the frequencyLabel, unitsLabel, lastValue, and footnote.
}
