% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_pdf.R
\name{chartLocation}
\alias{chartLocation}
\title{Set Graphical Parameters}
\usage{
chartLocation(
  location = "",
  margin = c(4, 2, 3, 3),
  x0 = NULL,
  x1 = NULL,
  y0 = NULL,
  y1 = NULL,
  newOverride = FALSE
)
}
\arguments{
\item{location}{Character variable. Six equal-sized graphs (three rows, two columns): top/middle/bottom, left/right (eg. topLeft, middleRight). Three equal-sized graphs (three rows): top/middle/bottom third (eg. topThird). Two equal-sized graphs (two rows) top/bottom half (eg. topHalf). Six unequal sized graphs (three rows, two columns, left-hand side 2/3rds, right-hand side 1/3rd): top/middle/bottom left2/right2 (eg. tL2, tL_big, tR_small). You can specify either in camelCase, lowercase, nn, nN, NN, lower case.}

\item{margin}{A numerical vector c(bottom, left, top, right). The default is c(4, 2, 3, 3).}

\item{newOverride}{Boolean variable which by default is FALSE. If TRUE and the location provided is not in the top left portion (eg. topLeft, topThird), it will clean the frame before drawing (aka start a new page). If you provided coordinates, FALSE will start a new page and TRUE will draw on the same page.}

\item{x0/y0}{coordinates of points *from* which to draw the chart. X is from left to right. Y is from bottom to top. Points are based off of the proportion of the PDF page. Eg. x0 = 0, x1 = 1/2 would be the left half of the page.}

\item{x1/y1}{coordinates of points *to* which to draw the chart.}
}
\description{
chartLocation specifies your chart location. Parameters can be set by specifying standard locations (eg. topLeft, topRight) or by specifying coordinates of points.
}
