% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaHelperFunctions.R
\name{colorSelector}
\alias{colorSelector}
\title{automatic color selector}
\usage{
colorSelector(numberOfSeries, colorScheme = "paulTol")
}
\description{
This selects color schemes based on the number of series
}
