% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datable.R
\name{datable}
\alias{datable}
\title{DaTable}
\usage{
datable(
  dataFrame,
  series,
  columnNames = NULL,
  chartTitle = "",
  footNote = "",
  marLeft = 0.01,
  marRight = 0.01,
  marTop = 0.01,
  marBottom = 0.01,
  shiftLeft = 0.05,
  shiftDown = 0,
  firstColShift = 0,
  firstRowShift = 0,
  colSpacing = NULL,
  fontSize = 0.5,
  colFontSize = 0.5,
  chartTitleFontSize = 0.85,
  colAlignment = 0,
  chartTitleAdj = 0,
  colorMatrix = NULL
)
}
\description{
This creates a data table
}
