% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_axis.R
\name{defineXAxisBar}
\alias{defineXAxisBar}
\title{defineXAxisBar}
\usage{
defineXAxisBar(
  df_full = plot.data.full,
  series1Length,
  series2Length,
  series3Length,
  bars,
  xBarTickAdj,
  xMajorTickFreq,
  xMinorTickFreq,
  xBarTicksSeries1,
  xBarTicksSeries2,
  xBarTicksSeries3,
  xSeriesDivider
)
}
\description{
This defines the X-axis for Bar Charts
}
