% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_axis.R
\name{defineXAxisDate}
\alias{defineXAxisDate}
\title{defineXAxisDate}
\usage{
defineXAxisDate(
  df = plot.data,
  xMajorTickFreq,
  xMinorTickFreq,
  xLabelCenter = T
)
}
\description{
This defines the X-axis for charts with a date x-axis.
}
