% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_axis.R
\name{defineXAxisNumeric}
\alias{defineXAxisNumeric}
\title{defineXAxisNumeric}
\usage{
defineXAxisNumeric(
  df = plot.data,
  xMajorTickFreq,
  xMinorTickFreq,
  roundDig,
  seriesX
)
}
\description{
This defines the X-axis for charts with a numeric x-axis
}
