% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_axis.R
\name{drawXAxis}
\alias{drawXAxis}
\title{drawXAxis}
\usage{
drawXAxis(xAxisFontSize = NULL, xLabelFreq, xTicksBelowAxis, xTicksLabels)
}
\description{
This draws the X-axis using the output from the defineXAxis functions.
}
