% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_axis.R
\name{drawYAxis}
\alias{drawYAxis}
\title{drawYAxis}
\usage{
drawYAxis(
  df = plot.data,
  dfLeft = NULL,
  yMin,
  yMax,
  yStep,
  yMinLeft,
  yMaxLeft,
  yStepLeft,
  yAxisFontSize,
  hadj,
  axisColorRight,
  axisColorLeft,
  roundDig,
  roundDigLeft,
  chartParametersRight,
  chartParametersLeft
)
}
\description{
This draws the Y-axis.
}
