% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaHelperFunctions.R
\name{eventLine}
\alias{eventLine}
\title{Eventline}
\usage{
eventLine(
  date = NA,
  lineHeight = NULL,
  textHeight = NULL,
  cex = opt$label.cex,
  adj = NULL,
  lwd = 1,
  oneLine = FALSE,
  label,
  lastFOMCdate = FALSE
)

eventLine(
  date = NA,
  lineHeight = NULL,
  textHeight = NULL,
  cex = opt$label.cex,
  adj = NULL,
  lwd = 1,
  oneLine = FALSE,
  label,
  lastFOMCdate = FALSE
)
}
\arguments{
\item{date}{}

\item{lineHeight}{Numeric. Height of line as percent of chart area.}

\item{textHeight}{Numeric. Vertical location of label as percent of chart area. If oneLine == FALSE, default is lineHeight + 0.05. If oneLine == TRUE, default is lineHeight + 0.09.}

\item{cex}{Default is 0.75.}

\item{adj}{Numeric vector which specify the x and y adjustment ('justification') of the label. With 0 for left/bottom and 1 for right/top. Default is 0.5 (centered).}

\item{lwd}{Numeric vector. Default is 1.}

\item{oneLine}{Boolean. Changes the default lineHeight and default distance between lineHeight and textHeight.}

\item{label}{Character vector or expression specifying the text to be written.}

\item{lastFOMCdate}{Boolean. Default is FALSE. If TRUE, uses stfm.helper::get_recent_fomc() to set the date.}

\item{pos}{Numeric. A position specifier for the label. Overrides any adj value. Values of 1, 2, 3, and 4, correspond to below, to the left of, above, and to the right of.}

\item{offset}{Numeric. If pos is specified, controls the distance of the text label from the specified coordinate in fractions of a character width.}

\item{lty}{Vector. Can be a mixture of numeric and character (eg. 1, "solid", 2, "dashed", 3, "dotted"). Default is 1 (solid).}
}
\description{
Creates a vertical line with label. It uses graphics::segments() and graphics::text()

This creates event lines inthe chart
}
