% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaHelperFunctions.R
\name{eventLineDate}
\alias{eventLineDate}
\title{Eventline}
\usage{
eventLineDate(
  date = fomcdates(-1),
  lineHeight = NULL,
  textHeight = NULL,
  cex = 0.4,
  adj = 0.5,
  lwd = 1,
  oneLine = FALSE,
  label
)
}
\description{
This does something else.
}
