% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_pdf.R
\name{exhibitTemplate}
\alias{exhibitTemplate}
\title{View an example exhibit function. For easy copy/paste.}
\usage{
exhibitTemplate()
}
\description{
exhibitTemplate shows a blank exhibit function to easily copy and paste it into your code.
}
