% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_labels.R
\name{flourish}
\alias{flourish}
\title{flourish}
\usage{
flourish(
  lineAtZero = F,
  df = NULL,
  recessions = F,
  customShading = F,
  shadeEnd = "",
  shadeStart = "",
  colors = NA,
  yMax = NA,
  seriesX = "date"
)
}
\description{
This applies recession/custom shading and lineAtZero.
}
