% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_labels.R
\name{getCopy}
\alias{getCopy}
\title{getCopy}
\usage{
getCopy(
  dataFrame = NULL,
  series = NULL,
  famePath = NULL,
  startDate = NULL,
  endDate = NULL,
  seriesScale = 1,
  interp = F,
  seriesX = "date"
)
}
\description{
This pulls the data from fame if necessary, subsets the data to the desired range, and applies any scaling/interpolation.
}
