% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_pdf.R
\name{getSecurityLabel}
\alias{getSecurityLabel}
\title{Get Security Label}
\usage{
getSecurityLabel(security, audience = FALSE, audienceType = 1)
}
\arguments{
\item{security}{Character variable. Must be one of "Class I", "Class II", "Class III", "Restricted Controlled", "Restricted", "Internal", or "None".}

\item{audience}{Boolean variable which by default is FALSE. If TRUE, you can add an audience (eg. FRSOnly, Secure External).}

\item{audienceType}{Specify 1, 2, or 3 for FRSOnly, Secure External, or External (respectively). Some security levels are only allowed a certain audienceType. A warning message will appear if you call an audienceType that is not compatible with your provided security level.}
}
\description{
getSecurityLabel is a helper function which makes a security label to your pdf page with the security and, if specified, audience. Creates the label based off Outlook's sensitivity labels.
}
