% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_pdf.R
\name{getUpperLeft}
\alias{getUpperLeft}
\title{Get Upper Left Page Label}
\usage{
getUpperLeft(
  securityLabel = securityLabel,
  audienceLabel = audienceLabel,
  pageNum = NULL
)
}
\arguments{
\item{securityLabel}{String provided by getSecurityLabel helper function.}

\item{audienceLabel}{String provided by getSecurityLabel helper function.}

\item{pageNum}{Specify the page number}
}
\description{
This is used in creating the security label on the PDF. It will appear in this format: 'securityLable' // 'audienceLabel' - Page 'pageNum'
}
