% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeledBarChart.R
\name{labeledBarChart}
\alias{labeledBarChart}
\title{Bar Chart With X-axis Labels}
\usage{
labeledBarChart(
  series = "",
  dataFrame = NA,
  colors = opt$colors,
  yMin = NULL,
  yMax = NULL,
  yStep = NULL,
  xLabels = "",
  xTicksLeft = 0.56,
  legendCol = 1,
  legendNames = "",
  legend = FALSE,
  chartTitle = "",
  chartTitleFontSize = opt$paneltitlecex,
  labelFontSize = opt$legend.cex,
  unitsLabel = "",
  unitsLabelLeft = "",
  frequencyLabel = "",
  axisColorRight = "black",
  axisColorLeft = "black",
  beside = FALSE,
  footNote = "",
  footNotePlacement = 1.2,
  footNoteFontSize = opt$foot.cex,
  seriesScale = 1,
  yMaxLab = yMax,
  xLabelsScale = 0.65,
  xLabelsPerp = TRUE,
  singleBarColor = TRUE,
  barBorder = T,
  barBorderColor = "black",
  barBorderWidth = 1,
  legendLocationX = 0.05,
  legendLocationY = 1,
  legendColPlacement = NULL,
  legendInset = 0.1,
  legendFontSize = opt$key.cex,
  dynamicColors = FALSE,
  centerFirstLabels = FALSE,
  palette = NULL,
  scheme = NULL
)
}
\description{
This creates a bar chart that has labels on the x-axis instead of dates
}
