% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_labels.R
\name{legend2}
\alias{legend2}
\title{legend2}
\usage{
legend2(
  x,
  y = NULL,
  legend,
  fill = NULL,
  col = par("col"),
  border = "black",
  lty,
  lwd,
  pch,
  angle = 45,
  density = NULL,
  bty = "o",
  bg = par("bg"),
  box.lwd = par("lwd"),
  box.lty = par("lty"),
  box.col = par("fg"),
  pt.bg = NA,
  cex = 1,
  pt.cex = cex,
  pt.lwd = lwd,
  xjust = 0,
  yjust = 1,
  x.intersp = 1,
  y.intersp = 1,
  adj = c(0, 0.5),
  text.width = NULL,
  text.col = par("col"),
  text.font = NULL,
  merge = do.lines && has.pch,
  trace = FALSE,
  plot = TRUE,
  ncol = 1,
  horiz = FALSE,
  title = NULL,
  inset = 0,
  xpd,
  title.col = text.col,
  title.adj = 0.5,
  seg.len = 2
)
}
\description{
Allows the columns in base R legend to have better widths.
}
