% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineChart.R
\name{lineChart}
\alias{lineChart}
\title{Line Chart}
\usage{
lineChart(
  dataFrame = NULL,
  dataFrameLeft = dataFrame,
  famePath = NULL,
  famePathLeft = famePath,
  series = NULL,
  seriesLeft = NULL,
  interp = FALSE,
  chartTitle = "",
  chartTitleFontSize = opt$paneltitlecex,
  seriesX = "date",
  seriesXLabel = "",
  startDate = NULL,
  endDate = NULL,
  xMajorTickFreq = "year",
  xMinorTickFreq = "month",
  xLabelFreq = "month",
  xLabelCenter = T,
  xTicksBelowAxis = F,
  extraLimX = 0,
  yMin = NULL,
  yMax = NULL,
  yStep = NULL,
  seriesScale = 1,
  unitsLabel = "",
  hadj = opt$yaxis.pos,
  roundDig = 2,
  yMinLeft = NULL,
  yMaxLeft = NULL,
  yStepLeft = NULL,
  seriesScaleLeft = 1,
  unitsLabelLeft = "",
  roundDigLeft = 2,
  axisColorRight = "black",
  axisColorLeft = "black",
  labelFontSize = opt$legend.cex,
  xAxisFontSize = opt$axis.cex,
  yAxisFontSize = opt$axis.cex,
  colorBandSeries = NA,
  seriesBandUpper = NA,
  seriesBandLower = NA,
  borderBand = F,
  legend = TRUE,
  legendNames = NULL,
  legendCol = 1,
  legendColPlacement = NULL,
  legendLocationX = 0.05,
  legendLocationY = 0.97,
  legendFontSize = opt$key.cex,
  footNote = "",
  footNotePlacement = NULL,
  footNoteFontSize = opt$foot.cex,
  printLastValue = TRUE,
  lastValueFreq = NA,
  lastValue = NULL,
  lastValueLocationX = 0.9,
  lastValueLocationY = 0.9,
  lastValuePrelim = F,
  frequencyLabel = "",
  colors = NULL,
  palette = NULL,
  scheme = NULL,
  dynamicColors = FALSE,
  lineAtZero = FALSE,
  recessions = FALSE,
  customShading = FALSE,
  shadeStart = "",
  shadeEnd = "",
  lineType = NULL,
  lineWidth = NULL,
  PRESET = list(),
  dataType = NULL,
  vertShift = NULL,
  horizShift = NULL
)
}
\arguments{
\item{dataFrame}{}

\item{dataFrameLeft}{}

\item{famePath}{}

\item{famePathLeft}{}

\item{series}{}

\item{seriesLeft}{}

\item{interp}{Boolean variable. Default is FALSE. If TRUE, replaces NA by interpolation using zoo::na.approx().}

\item{chartTitle}{Character variable.}

\item{chartTitleFontSize}{Default is 0.85.}

\item{seriesX}{Character vector. Default is c("date").}

\item{seriesXLabel}{Character variable. Typically used if seriesX != c("date").}

\item{startDate}{If seriesX == c("date"), character variable. If is.numeric(seriesX), numberic variable.}

\item{endDate}{}

\item{xMajorTickFreq}{If seriesX == c("date"), character variable (eg. "year", "quarter", "month", "day").}

\item{xMinorTickFreq}{}

\item{xLabelFreq}{}

\item{xLabelCenter}{Boolean variable. Default is TRUE.}

\item{xTicksBelowAxis}{Boolean variable. Default is FALSE.}

\item{extraLimX}{Numeric variable. Adjusts x-axis limits.}

\item{yMin}{}

\item{yMax}{}

\item{yStep}{}

\item{seriesScale}{}

\item{unitsLabel}{Character variable.}

\item{hadj}{Numeric variable. Default is 0.5. Justification for all labels parallel to reading direction.}

\item{roundDig}{Integer indicating the number of decimal places to be used. Default is 2.}

\item{yMinLeft}{}

\item{yMaxLeft}{}

\item{yStepLeft}{}

\item{seriesScaleLeft}{}

\item{unitsLabelLeft}{}

\item{roundDigLeft}{Integer indicating the number of decimal places to be used. Default is 2.}

\item{axisColorRight}{Character variable. Default is "black".}

\item{axisColorLeft}{}

\item{labelFontSize}{Default is 0.7.}

\item{xAxisFontSize}{Default is 0.65.}

\item{yAxisFontSize}{Default is 0.65.}

\item{colorBandSeries}{Character variable. Color of banded series. Default is "#0000FF33".}

\item{seriesBandUpper}{Character vector. Series used to be upper line of polygon.}

\item{seriesBandLower}{Character vector. Series used to be lower line of polygon.}

\item{borderBand}{The color to draw the border. The default, NULL, means to use par("fg"). Use borderBand = NA to omit borders.}

\item{legend}{Boolean variable. Default is TRUE.}

\item{legendNames}{Character vector.}

\item{legendCol}{Integer. Default is 1. Number of columns in legend.}

\item{legendColPlacement}{Numeric. Must be >= 0. Allows the columns to have better widths.}

\item{legendLocationX}{Numeric. Horizontal location of legend as percent of chart area. Default is 0.05.}

\item{legendLocationY}{Numeric. Vertical location of legend as percent of chart area. Default is 0.97.}

\item{legendFontSize}{Default is 0.75.}

\item{footNote}{Character vector. c("Note:  Best practice to make the footNote all one vector (for indentation), to double space after 'note' and 'source', and to have 'note' first and then 'source.'\ n Source:  PolicyPlot charting dos and donts.")}

\item{footNotePlacement}{Numeric. on which margin line, starting at 0 counting outwards.}

\item{footNoteFontSize}{Default is 0.55.}

\item{printLastValue}{Boolean. Default is TRUE.}

\item{lastValueFreq}{If seriesX == c("date"), character variable (eg. "year", "quarter", "month", "day").}

\item{lastValue}{Specify lastValue}

\item{lastValueLocationX}{Numeric. Horizontal location of last value as percent of chart area. Default is 0.9.}

\item{lastValueLocationY}{Numeric. Vertical location of last value as percent of chart area. Default is 0.9.}

\item{lastValuePrelim}{Boolean. Default is FALSE. If TRUE, adds * after last value.}

\item{frequencyLabel}{Character variable (eg. "year", "quarter", "month", "day").}

\item{colors}{Character vector. Defaults to opt$colors.}

\item{palette}{Character variable (eg. "s", "d", "q" / "sequential", "diverging", "qualitative").}

\item{scheme}{Character variable. See pickColors() in helper_colors.R for complete list.}

\item{dynamicColors}{Boolean. Default is FALSE. If TRUE, uses PaulTol color scheme.}

\item{lineAtZero}{Boolean. Default is FALSE. If TRUE, adds horizontal, solid, black line at y = 0.}

\item{recessions}{Boolean. Default is FALSE. If TRUE, shades recessions using fiafunctions::nber.shade().}

\item{customShading}{Boolean. Default is FALSE. If TRUE, shades region using fiafunctions::shadedRegion().}

\item{shadeStart}{}

\item{shadeEnd}{}

\item{lineType}{Vector. Can be a mixture of numeric and character (eg. 1, "solid", 2, "dashed", 3, "dotted").}

\item{lineWidth}{Numeric vector. fiafunctions default is 1.7.}

\item{PRESET}{list(), used for common chart set ups. Variables within the list will override the default values, unless overridden themselves by variables explicitly specified by the user in the chart call function.}

\item{dataType}{Depreciated. Used to signify if data was from an R object or FAME.}

\item{vertShift}{Depreciated. Replaced with lastValuePlacementY.}

\item{horizShift}{Depreciated. Replaced with lastValuePlacementX.}
}
\description{
This creates a line chart.
}
\examples{
\dontrun{
lineChart(chartTitle="Example Line Chart",
         legendNames=c("Value 1", "Value 2"),
         series=c("value", "value2"),
         dataFrame=testDataFrame,
         frequencyLabel="Decribe Legend Here",
         #yStep=2,
         #yMin=-2,
         #yMax=40,
         footNote=c('Insert Footnote Text'),
         legendCol=1,
         startDate="2006-01-01",
         unitsLabel="Units of Data",
         footNotePlacement=1.5,
         colors = c("firebrick","steelblue3"),
         dataType="r"
         )
}
         
}
