% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_labels.R
\name{lineFormatting}
\alias{lineFormatting}
\title{lineFormatting}
\usage{
lineFormatting(
  lineType = NULL,
  lineWidth = NULL,
  series = NULL,
  seriesLeft = NULL
)
}
\description{
This assigns line types and line widths if none are supplied.
}
