% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_pdf.R
\name{newPage}
\alias{newPage}
\title{Set PDF Parameters}
\usage{
newPage()
}
\description{
newPage can be used to set FIA's standard PDF parameters. It sets the outer margins (omi) as (bottom = 0.275", left = 0.2", top = 0.65", right = 0.5") and the margin size (mai) as (bottom = 3.5", left = 3", top = 3.5", right = 3").
}
