% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_pdf.R
\name{pageLabels}
\alias{pageLabels}
\title{Create page labels.}
\usage{
pageLabels(
  exhibit = NULL,
  security,
  audience = FALSE,
  audienceType = 1,
  pageNum = NULL,
  timeStamp = TRUE,
  footNoteFIA = TRUE,
  footNotePage = NULL
)
}
\arguments{
\item{exhibit}{Default is NULL. If specified, will add the exhibit name to the top center margin of the page.}

\item{security}{Character variable. Must be one of "Class I", "Class II", "Class III", "Restricted Controlled", "Restricted", "Internal", or "None".}

\item{audience}{Boolean variable which by default is FALSE. If TRUE, you can add an audience (eg. FRSOnly, Secure External).}

\item{audienceType}{Specify 1, 2, or 3 for FRSOnly, Secure External, or External (respectively). Some security levels are only allowed a certain audienceType. A warning message will appear if you call an audienceType that is not compatible with your provided security level.}

\item{pageNum}{Specify the page number}

\item{timeStamp}{Boolean variable which by default is TRUE. If FALSE, will not add a date and time stamp on upper right margin}

\item{footNoteFIA}{Boolean variable which by default is TRUE. If FALSE, will not add MA - Financial Intermediaries Analysis Section on bottom margin}

\item{footNotePage}{Default is NULL. If specified, will add the specified string to the bottom margin of the page above any FIA footnote that might be there. It will be center justified and the font size is 0.6.}
}
\description{
pageLabels creates your page labels: exhibit name, security, audience, page number, add a time stamp, add FIA on the bottom, add a footnote.
}
