% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_colors.R
\name{pickColors}
\alias{pickColors}
\title{Pick Colors}
\usage{
pickColors(
  series = NULL,
  seriesLeft = NULL,
  colors = NULL,
  dynamicColors = FALSE,
  palette = NULL,
  scheme = NULL
)
}
\description{
This assigns colors.
}
