% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterChart.R
\name{scatterChart}
\alias{scatterChart}
\title{Scatter Chart}
\usage{
scatterChart(
  dataFrame = NULL,
  famePath = NULL,
  series = NULL,
  chartTitle = "",
  chartTitleFontSize = opt$paneltitlecex,
  seriesX = "date",
  seriesXLabel = "",
  startDate = NULL,
  endDate = NULL,
  xMajorTickFreq = "year",
  xMinorTickFreq = "month",
  xLabelFreq = "month",
  xTicksBelowAxis = F,
  extraLimX = 0,
  yMin = NULL,
  yMax = NULL,
  yStep = NULL,
  seriesScale = 1,
  unitsLabel = "",
  hadj = opt$yaxis.pos,
  roundDig = 2,
  axisColorRight = "black",
  axisColorLeft = "black",
  labelFontSize = opt$legend.cex,
  xAxisFontSize = opt$axis.cex,
  yAxisFontSize = opt$axis.cex,
  legend = TRUE,
  legendNames = NULL,
  legendCol = 1,
  legendColPlacement = strwidth(max(legendNames)),
  legendLocationX = 0.05,
  legendLocationY = 0.97,
  legendFontSize = opt$key.cex,
  footNote = "",
  footNotePlacement = NULL,
  footNoteFontSize = opt$foot.cex,
  printLastValue = FALSE,
  lastValueFreq = NA,
  lastValue = NULL,
  lastValueLocationX = 0.9,
  lastValueLocationY = 0.9,
  frequencyLabel = "",
  colors = NULL,
  palette = NULL,
  scheme = NULL,
  dynamicColors = FALSE,
  lineAtZero = FALSE,
  recessions = FALSE,
  customShading = FALSE,
  shadeStart = "",
  shadeEnd = "",
  pointType = NULL,
  trendLine = F,
  lineType = NULL,
  lineWidth = NULL,
  dataType = NULL,
  horizShift = NULL,
  vertShift = NULL
)
}
\arguments{
\item{dataFrame}{}

\item{famePath}{}

\item{series}{}

\item{chartTitle}{Character variable.}

\item{chartTitleFontSize}{Default is 0.85.}

\item{seriesX}{Character vector. Default is c("date").}

\item{seriesXLabel}{Character variable. Typically used if seriesX != c("date").}

\item{startDate}{If seriesX == c("date"), character variable. If is.numeric(seriesX), numberic variable.}

\item{endDate}{}

\item{xMajorTickFreq}{If seriesX == c("date"), character variable (eg. "year", "quarter", "month", "day").}

\item{xMinorTickFreq}{}

\item{xLabelFreq}{}

\item{xTicksBelowAxis}{Boolean variable. Default is FALSE.}

\item{extraLimX}{Numeric variable. Adjusts x-axis limits.}

\item{yMin}{}

\item{yMax}{}

\item{yStep}{}

\item{seriesScale}{}

\item{unitsLabel}{Character variable.}

\item{hadj}{Numeric variable. Default is 0.5. Justification for all labels parallel to reading direction.}

\item{roundDig}{Integer indicating the number of decimal places to be used. Default is 2.}

\item{axisColorRight}{Character variable. Default is "black".}

\item{axisColorLeft}{}

\item{labelFontSize}{Default is 0.7.}

\item{xAxisFontSize}{Default is 0.65.}

\item{yAxisFontSize}{Default is 0.65.}

\item{legend}{Boolean variable. Default is TRUE.}

\item{legendNames}{Character vector.}

\item{legendCol}{Integer. Default is 1. Number of columns in legend.}

\item{legendColPlacement}{Numeric. Must be >= 0. Allows the columns to have better widths.}

\item{legendLocationX}{Numeric. Horizontal location of legend as percent of chart area. Default is 0.05.}

\item{legendLocationY}{Numeric. Vertical location of legend as percent of chart area. Default is 0.97.}

\item{legendFontSize}{Default is 0.75.}

\item{footNote}{Character vector. c("Note:  Best practice to make the footNote all one vector (for indentation), to double space after 'note' and 'source', and to have 'note' first and then 'source.'\ n Source:  PolicyPlot charting dos and donts.")}

\item{footNotePlacement}{Numeric. on which margin line, starting at 0 counting outwards.}

\item{footNoteFontSize}{Default is 0.55.}

\item{printLastValue}{Boolean. Default is FALSE.}

\item{lastValueFreq}{If seriesX == c("date"), character variable (eg. "year", "quarter", "month", "day").}

\item{lastValue}{Specify lastValue}

\item{lastValueLocationX}{Numeric. Horizontal location of last value as percent of chart area. Default is 0.9.}

\item{lastValueLocationY}{Numeric. Vertical location of last value as percent of chart area. Default is 0.9.}

\item{frequencyLabel}{Character variable (eg. "year", "quarter", "month", "day").}

\item{colors}{Character vector. Defaults to opt$colors.}

\item{palette}{Character variable (eg. "s", "d", "q" / "sequential", "diverging", "qualitative").}

\item{scheme}{Character variable. See pickColors() in helper_colors.R for complete list.}

\item{dynamicColors}{Boolean. Default is FALSE. If TRUE, uses PaulTol color scheme.}

\item{lineAtZero}{Boolean. Default is FALSE. If TRUE, adds horizontal, solid, black line at y = 0.}

\item{recessions}{Boolean. Default is FALSE. If TRUE, shades recessions using fiafunctions::nber.shade().}

\item{customShading}{Boolean. Default is FALSE. If TRUE, shades region using fiafunctions::shadedRegion().}

\item{shadeStart}{}

\item{shadeEnd}{}

\item{pointType}{Numeric vector. Defaults tp pch = 20.}

\item{trendLine}{Boolean variable. Default is FALSE.}

\item{lineType}{Vector. Can be a mixture of numeric and character (eg. 1, "solid", 2, "dashed", 3, "dotted").}

\item{lineWidth}{Numeric vector. fiafunctions default is 1.7.}

\item{dataType}{Depreciated. Used to signify if data was from an R object or FAME.}

\item{horizShift}{Depreciated. Replaced with lastValuePlacementX.}

\item{vertShift}{Depreciated. Replaced with lastValuePlacementY.}
}
\description{
This creates a scatter chart.
}
\examples{
\dontrun{
scatterChart(
     chartTitle = 'Sepal width vs Sepal Length',
     dataFrame = chart_data,
     seriesX = c('Sepal.Length'), # given if not using the default date for x axis, must be numeric
     series = c("setosa", "versicolor", "virginica"),
     legendNames = c("setosa", "versicolor", "virginica"),
     dynamicColors = T,
     yMax = 4.5,
     yMin = 2,
     yStep = 0.5,
     startDate = 4,
     endDate = 8,
     xMajorTickFreq = 0.5,
     xMinorTickFreq = .25,
     customShading = T,
     shadeStart = 5,
     shadeEnd = 7,
     trendLines = T,
   )
   }
   
}
