% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaHelperFunctions.R
\name{set_chart_parameters}
\alias{set_chart_parameters}
\title{chart parameters}
\usage{
set_chart_parameters(
  ...,
  envir = parent.frame(),
  set_colors = TRUE,
  set_sizes = TRUE
)
}
\description{
sets chart's parameters
}
