% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_labels.R
\name{setYLims}
\alias{setYLims}
\title{setYLims}
\usage{
setYLims(
  df = NULL,
  dfLeft = NULL,
  series = NULL,
  seriesLeft = NULL,
  yMax = NULL,
  yMin = NULL,
  yStep = NULL,
  yMaxLeft = NULL,
  yMinLeft = NULL,
  yStepLeft = NULL
)
}
\description{
This sets the Y-axis limits for the chart given the users inputted yMax, yMin, and yStep parameters. This is done for both the right and left axis at the same time.
}
