% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaHelperFunctions.R
\name{shadedRegion}
\alias{shadedRegion}
\title{General Shading}
\usage{
shadedRegion(
  plotteddates,
  shadeyval,
  shadeStartDate,
  shadeEndDate,
  shadecol = "gray88",
  shadeColBorder = "gray88",
  bottom = -10000
)
}
\description{
This shades recessions
}
